requirements.txt
streamlit
fastapi
uvicorn
faiss-cpu
pandas
langchain
langchain-community
langchain-ollama
ollama
python-multipart
torchvision